requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dnspython==2.7.0
docopt==0.6.2
email_validator==2.2.0
fastapi==0.115.5
fastapi-cli==0.0.5
Flask==3.1.0
Flask-GraphQL==2.0.1
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
graphql-server-core==1.2.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
pipreqs==0.4.13
promise==2.3
pydantic==2.10.1
pydantic_core==2.27.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
Rx==1.6.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
strawberry-graphql==0.251.0
typer==0.13.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==0.24.0
websockets==14.1
Werkzeug==3.1.3
yarg==0.1.10
